requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
pydantic==2.12.3
pydantic-settings==2.0.3
asyncpg==0.30.0
alembic==1.12.1
python-dotenv==1.0.0
bcrypt==4.2.0
PyJWT==2.8.0
python-multipart==0.0.9
cryptography==43.0.3